Create IOT Based SMoke Alarm SYstem + Blynk App + Google Firebase
This project is a smart IoT-based smoke and gas detector using NodeMCU and the MQ-2 sensor. It continuously monitors levels of smoke, LPG, and carbon monoxide, triggering a buzzer and LED alarm when dangerous levels are detected. The system also sends real-time data and alerts via Blynk and logs every reading to Firebase with timestamps, creating a reliable record for monitoring and analysis. It’s a compact yet powerful solution that combines safety, connectivity, and cloud intelligence in one